pyproject.toml
[project]
name = "arfbot"
version = "0.1.0"
description = "Simple telegram football game bot"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.24.0",
    "aiosqlite>=0.22.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
]
